# internal/spec/openapi.yaml
openapi: 3.1.0
info:
  title: hello
  version: 1.1.0
  description: Wake-on-LAN registry with live presence. Send magic packets, track
    devices, and subscribe to status changes over WebSocket.
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /monitor/check:
    post:
      tags: [Monitor]
      summary: Poll every device now
      operationId: checkMonitor
      responses:
        '200':
          description: Immediate polling pass outcome
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /monitor/transitions:
    get:
      tags: [Monitor]
      summary: Latest reachability changes
      operationId: monitorTransitions
      responses:
        '200':
          description: Up to 20 transitions, newest first
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransitionsResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /monitor/uptime:
    get:
      tags: [Monitor]
      summary: Per-device poll uptime
      operationId: monitorUptime
      responses:
        '200':
          description: Up/total poll counts since process start
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UptimeResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /monitor/status:
    get:
      tags: [Monitor]
      summary: Monitor tuning and latest pass
      operationId: monitorStatus
      responses:
        '200':
          description: Interval, timeout, and latest pass counters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonitorStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
...
  schemas:
...
    CheckResult:
      type: object
      required: [checked, changed]
      properties:
        checked:
          type: integer
        changed:
          type: integer
      additionalProperties: false
    Transition:
      type: object
      required: [device_id, from, to, at]
      properties:
        device_id:
          type: string
        from:
          type: string
        to:
          type: string
        at:
          type: integer
          format: int64
      additionalProperties: false
    TransitionsResult:
      type: object
      required: [transitions]
      properties:
        transitions:
          type: array
          items:
            $ref: '#/components/schemas/Transition'
      additionalProperties: false
    UptimeEntry:
      type: object
      required: [device_id, up, total, percent]
      properties:
        device_id:
          type: string
        up:
          type: integer
        total:
          type: integer
        percent:
          type: number
          description: 100 * up / total
      additionalProperties: false
    UptimeResult:
      type: object
      required: [devices]
      properties:
        devices:
          type: array
          items:
            $ref: '#/components/schemas/UptimeEntry'
      additionalProperties: false
...
    MonitorStatus:
      type: object
      required: [interval_seconds, timeout_seconds, last_run, last_checked, last_changed]
      properties:
        interval_seconds:
          type: integer
        timeout_seconds:
          type: integer
        last_run:
          type: integer
          format: int64
          description: Unix timestamp of the latest pass, 0 if none
        last_checked:
          type: integer
        last_changed:
          type: integer
      additionalProperties: false
...
    Problem:
      type: object
      description: RFC 9457 problem details. All errors share this envelope.
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          format: uri
          example: about:blank
        title:
          type: string
          example: not found
        status:
          type: integer
          example: 404
        detail:
          type: string
          example: 'storage: device not found'
      additionalProperties: false
...
  responses:
...
    Unauthorized:
      description: Missing or invalid bearer token, RFC 9457 body
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API token from API_TOKEN. Browsers calling /ws use ?token= instead.